// src/store/pagination/types.ts
import { Action } from 'redux';
import { MovieType } from '../ui/types';

export interface Movie {
  poster_path?: string | null;
  adult?: boolean;
  overview?: string;
  release_date?: string;
  genre_ids?: number[];
  id?: number;
  original_title?: string;
  original_language?: string;
  title?: string;
  backdrop_path?: string | null;
  popularity?: number;
  vote_count?: number;
  video?: boolean;
  vote_average?: number;
}

export type MovieId = Movie['id'];

export interface PaginationState {
  page: number;
  ids: MovieId[];
  loading: boolean;
  totalPages?: number;
  totalResults?: number;
}

export type PaginationActionType = string;

export type EntitiyType<T> = { [id: number]: T };

export type PaginationPaylod = {
  result: PaginationState;
  entities: { movies: EntitiyType<Movie> };
};

export interface PaginationAction extends Action<PaginationActionType> {
  movieType: MovieType;
  payload?: PaginationPaylod;
  url?: string;
  page?: number;
  query?: string;
}